Fix Person::isSamePerson and add tests to reflect the change.
Originally, isSamePerson would return true if a person has the same name, and same email or phone number. However, if two people had the same name and no phone/no email, this "no phone/email" would be considered the same phone/email, and trigger true. This commit fixes this, and adds tests to ensure that.