Update CompleteCommand and CompleteCommandParser
The CompleteCommand relied on a Set of indexes in order to subsequently
parse all of them. This is not feasible as it requires the parser to
fetch a set of Indexes, which it cannot as it does not have access to
the filteredList of Tasks nor the model.

Thus, the responsibility of finding the appropriate Tasks has been
propogated downwards to the CompleteCommand class

Now the CompleteCommand class handles the fetching of tasks for
predicate based entry. I.e. given a Task<Predicate> completeCommand will
be able to complete a set of completable tasks fulfilling that
predicate.

- Support for predicate based completion in the CompleteCommandParser
comes in the form of labels now being parsed to create predicates that
returns true on a case-senstive match with the tasks