Refactor AddCommand to PlaylistNewCommand
Since we are translating Person to Playlist,
the original `add` is translated to `playlist new`

The old AddressBook argument prefixes in CliSyntax are removed.
Those prefixes are only used in codes relating to AddCommand and EditCommand.
ie AddCommand, AddCommandParser, and their tests, and the EditCommand counterparts
Since codes relating to EditCommand is commented out for reference,
it wasn't fully refactored.
So only codes relating to AddCommand are fully refactored.