Change validation for Track
As Track is translated from Tag, the original validation method is the isValidTag method.
The method was adapted to a isValidMp3 stub in PR #42.

Now the Track constructors is used as the entry point of validation.
The constructors validates for file inexistence and javafx media support
using the AppUtil.checkArgument method, so it throws IllegalArgumentException.

Todo: move validation for javafx media support to PlayableTrack