Fix Travis build error
Issue was due to file name case sensitivity.
With exceptions stack trace enabled in travis logs,
I found out that Travis could not locate "library/Ihojin no Yaiba.mp3".
https://travis-ci.org/CS2103-AY1819S1-T13-3/main/builds/443999074
I noticed that the actual file name is in lower case,
while the file name in CommandTestUtil uses the capitalized name.

Thus the fix was either to change CommandTestUtil to use lower case
or to rename the file to capitalized format.
I choose to rename the file.
Also, git is case insensitive, so I cannot use "git add" to stage
but needed to use "git mv" according to
https://stackoverflow.com/a/24979063/4858751

File names in ParserUtilTest and PlaylistCardTest are also fixed.