Fix bug in test where points are not updating
Note: There was a revert for this fix earlier as it included changes to
the project that was not intended. i.e. accidental `git add .`

Previously, a bug is encountered where the expectedMessage will
report change in points as zero.

This is because procedural style code has to be used. The
oldXP needs to be obtained from the model before updates to the
model, and the newXp needs to be obtained after the change. but due to
a bug where the newXp is obtained before the model is changed, the
reported change is zero.

In the fix, the newXp is obtained only after model has changed.