Refactor PlayerManager to be a singleton
As Player will only be used in Command subclasses,
I converted the PlayerManager to be a singleton.
The singleton is only used in the Command abstract class.
This way, all subclasses will have access to the player instance.

An alternative would be to add a Player parameter to
every Command subclasses similar to how CommandHistory
used to be a parameter before the refactoring stage.
The argument for not going with this alternative is that
not all Command subclass need to use the Player,
just like CommandHistory was only used by undo and redo.
Also, I prefer to avoid further mass refactoring nightmare.

Of course, we need to be mindful of the consequences of
using singleton as singleton class is essentially a global variable.
We try to use it sparingly.
Pretty sure this is the only time where we need singleton.

Honestly, the initial codes of the AddressBook was not very
intuitive/friendly for integrating new components,
or dependency injection.