Change to to v1.3 in main app
Add new NoSelectionRequestEvent.java to handle deselect of car park card.
Add newline for DataFetchExceptionEvent.java
Add check for CarparkFinderParser.java where if the command has no arguements it should throw ParseException.
Remove unnccessary imports in GsonUtil.java, QueryCommand.java

Add new method isNonNegativeAboveNineInteger(). This helps to check if seconds for NotifyCommand is within the range of 10 to 600 and also allow 0 to turn it off.
Add post event in ClearCommand to deselect the car park card when there are no more car parks.
Add newline in FindCommand so that the MESSAGE_USAGE is not that long.
Move the logic for NotifyCommand to the new NotifyTimeTask.java, which will create a separate thread to handle the timer task.
Change the check for parseTime to NonNegativeAboveNineInteger() method in ParserUtil.java
NotifyCommand and QueryCommand now works using multithreading.
Add Timer and NoSelectionRequestEvent to CarparkListPanel.java
Enablle the Command box when exception is done
Make the Message Box bigger