MainApp: Load default data instead of empty
When the xml failed to parsed due to invalid field or if it cannot be read properly, a new file should be generated. Currently, it generates an empty file which means that the system will not be usable as it contains no default account.

Lets fix that by generating an empty file with a default `root` account instead.