LogicManager: Update subset of menu commands as guest commands
Currently, the application only allows the user to view the menu and navigate by using a mouse to click on the cards when the user is not authenticated. Since we are constrainted by CLI, this breaks the intended behavior if the user doesn't possess a mouse.

Thus, let's include the commands of Menu which will not modify the data as guest commands so users will be able to execute them without being logged in.