LogicManager: Remove sort command
Currently, guest user can execute the sort command.

As the sort command modifies the arrangement of the data, this has become inconsistent with the main intention of not being able to manipulate the data while using as a guest, which includes not modifying its positions.

Let's remove the command from the set of guest commands.