SampleDataUtil: Add method to get sample data
Currently, the sample data only contains the root account and nothing else.

This might be an issue when it comes to PE or product demonstration.

Let's update the file by including a method that populates the storage with decent amount of data. In addition, update related test cases (mainly the Clear-related cases) to use the method that generates only the root account and not the sample data.